requirements.txt
streamlit==1.24.0
transformers
torch==2.2.1
torchaudio==2.2.1
datasets==2.13.1
soundfile==0.12.1
python-dotenv==1.0.0
sentencepiece
python-docx==0.8.11
PyPDF2==3.0.1